# backend/openapischema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Document Analysis API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/insights/generate": {
      "get": {
        "tags": ["Insights"],
        "summary": "Generate and retrieve insights for an entity",
        "description": "Generates and retrieves structured insights for a specified document, collection, or recommendation. Exactly one of `doc_id`, `col_id`, or `rec_id` must be provided. The generated insight is stored and returned.",
        "operationId": "generate_insights_for_entity_api_v1_insights_generate_get",
        "parameters": [
          {
            "name": "doc_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ID of the document to generate insights for.",
              "title": "Doc Id"
            },
            "description": "ID of the document to generate insights for."
          },
          {
            "name": "col_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ID of the collection to generate insights for.",
              "title": "Col Id"
            },
            "description": "ID of the collection to generate insights for."
          },
          {
            "name": "rec_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ID of the recommendation to generate insights for.",
              "title": "Rec Id"
            },
            "description": "ID of the recommendation to generate insights for."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InsightInDB"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/insights/{insight_id}": {
      "get": {
        "tags": ["Insights"],
        "summary": "Retrieve a single insight by ID",
        "description": "Fetches a single insight's details using its unique identifier. Returns a 404 error if the insight is not found.",
        "operationId": "get_insight_by_id_api_v1_insights__insight_id__get",
        "parameters": [
          {
            "name": "insight_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The unique identifier of the insight to retrieve.",
              "title": "Insight Id"
            },
            "description": "The unique identifier of the insight to retrieve."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InsightInDB"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InsightInDB": {
        "properties": {
          "insightId": {
            "type": "string",
            "title": "Insightid",
            "description": "The unique identifier for the insight."
          },
          "sourceType": {
            "type": "string",
            "title": "Sourcetype",
            "description": "The type of the source ('document', 'collection', or 'recommendation')."
          },
          "sourceId": {
            "type": "string",
            "title": "Sourceid",
            "description": "The ID of the source entity."
          },
          "insights_data": {
            "items": {
              "$ref": "#/components/schemas/InsightItem"
            },
            "type": "array",
            "title": "Insights Data",
            "description": "The list of structured insight items."
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Generatedat",
            "description": "The timestamp when the insight was generated."
          }
        },
        "type": "object",
        "required": [
          "insightId",
          "sourceType",
          "sourceId",
          "insights_data",
          "generatedAt"
        ],
        "title": "InsightInDB",
        "description": "Pydantic model representing an insight as stored in the database.\n\nThis model is used for API responses.",
        "example": {
          "generatedAt": "2025-01-01T12:00:00",
          "insightId": "insight_abc123",
          "insights_data": [
            {
              "data": "The main point of the document is...",
              "priority": 1,
              "type": "key_insight"
            },
            {
              "data": "Did you know that...",
              "priority": 2,
              "type": "did_you_know"
            }
          ],
          "sourceId": "doc_xyz789",
          "sourceType": "document"
        }
      },
      "InsightItem": {
        "properties": {
          "type": {
            "$ref": "#/components/schemas/InsightType",
            "description": "The category of the insight."
          },
          "data": {
            "type": "string",
            "title": "Data",
            "description": "The actual text content of the structured insight."
          },
          "priority": {
            "type": "integer",
            "title": "Priority",
            "description": "A numerical priority for the insight, lower is higher priority.",
            "default": 1
          }
        },
        "type": "object",
        "required": ["type", "data"],
        "title": "InsightItem",
        "description": "Pydantic model for a single structured insight item."
      },
      "InsightType": {
        "type": "string",
        "enum": [
          "Key insights",
          "Did you know?",
          "Contradictions / counterpoints",
          "Inspirations or connections across docs",
          "generation_error"
        ],
        "title": "InsightType"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}